Added working UI flow for creatiing post templates and post.
This is a definitive commit as it connects most of the backend functions to the frontend side. While it is still in progress, it introduces UI for creating and viewing posts and post templates in a community. It currently needs some cleaning, addition of a third party api and some additional changes for later on.