Added unit tests for Post and PostTemplate models.
I added several unit test that checks validity of Post and PostTemplate models. I wrote this tests in another branch. They are working as expected but this branch had conficts with my test branch so I decided to commit to this branch directly. We can solve this branchs conflict with the base MainApp branch later on.