Model Update and Register/Login Functions
I omitted and renamed some of the Model fields for brewity and replaced our custom User model with Django's built-in User model . I also added a register and login function that utilizes Django authorization system. I did not wrote unit tests but I plan to write them as soon as possible.
There are also some functions for the home page but they are currently drafts so I did not include them in our URL routes.