Made some changes on GetCommunityData endpoint and added unit tests.
I made some tweaks on the code. Now community objects return with isJoined boolean field, which indicates whether the current user is joined to the community or not. I also added Posts field on GetCommunityData endpoint which automatically returns posts in the community.
I also added unit tests which are all passing except list community and list post template tests. I suspect it's related to a bug in the test, not in the actual code.
I also cleaned some of the older code left as comments and  removed some unnecessary fields in the serializers.