Added Postman Collection and tags to Emrah's contributions
I added a Postman Collection that includes API calls to CreateCommunity,CreatePost and CreatePostTemplate. I also added tags to the schemas to organize Swagger page. Note that descriptions inserted are provided by Emrah Doğan and pushed to the Backend branch before this PR. This was made to resolve conflicts.