Implented core funtionality of FilterPosts endpoint and added Swagger documentation.
I wrote a dynamic script that is designed to handle and apply arbitrary query functions to a given list of post objects. It expects query functions as lambda functions that are organized by their respective fields on the QueryFunctions class. It has some issues addressing URL unsafe field names and does not contain error handling mechanisms but I aim to address these issues later on.
I also filled in the description for the Swagger documentation that clearly explains its usage.